Create ResourceCheckerandVMStatusChanger
This PowerShell script primarily manages Azure resources and virtual machines. Here's a high-level summary:

Import necessary modules: The script starts by importing the necessary Azure modules (Az.Accounts, Az.Compute, Az.Resources) that will be used to interact with Azure resources.

Login to Azure account: The Login-AzAccount command initiates a login to Azure. A login screen will be prompted, unless you're already logged in.

Function definitions: The script defines a few functions to manage resources and VMs:

Get-ResourcesBySubscription: This function retrieves and lists all Azure resources under a given subscription ID. It includes VM statuses, AKS cluster statuses, and the usage of VMs in the 'East US' location.

Start-VMWithConfirm: This function starts a specified VM in a specified resource group, but only after the user confirms that they want to proceed.

Stop-VMWithConfirm: This function stops a specified VM in a specified resource group, but only after the user confirms that they want to proceed.

Main script: The main part of the script prompts the user to choose from one of several actions:

Check resources by subscription (1), resource group (2), or tag (3). When one of these options is selected, the user will be prompted to enter the relevant information (like the subscription ID), and the appropriate function will be called to retrieve and display the resources.

Start a VM (4) or stop a VM (5). When one of these options is selected, the user will be prompted to enter the resource group name and VM name, and the appropriate function will be called to start or stop the VM, with user confirmation.

NOTE: List VMs (6). THIS IS NOT DEFINED YET! 

If the user enters any other number, the script will print 'Invalid option selected'.